+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## 0.5.0 - TBD
+- Added support for Statistics schema
 - Changed `schema` and `stype_in` to optional in `Metadata` to support live data
 - Added `RType` enum for exhaustive pattern matching
 - Added `&str` getters for more `c_char` array record fields
 - Changed `DbnDecoder.decode` to always return a list of tuples
+- Fixed value associated with `Side::None`
 - Fixed issue with decoding partial records in Python `DbnDecoder`
 - Fixed missing type hint for Metadata bytes support